feat(delete): adiciona validacao de delete

+++ backend-with-ca/src/usecases/delete-todo/delete-todo.ts
@@ -1,4 +1,5 @@
-import { type Either, right } from '../../shared/either'
+import { type Either, right, left } from '../../shared/either'
+import { TodoNotFoundError } from '../create-new-todo/errors/todo-not-found-error'
 import { type ITodoRepository } from '../shared/ports/todo-repository'
 import { type IUseCase } from '../shared/ports/use-case'
 
@@ -8,7 +9,11 @@ class DeleteTodoUseCase implements IUseCase {
     this.todoRepository = todoRepository
   }
 
-  public async execute (todoId: string): Promise<Either<void, string>> {
+  public async execute (todoId: string): Promise<Either<TodoNotFoundError, string>> {
+    const todoExists = await this.todoRepository.findById(todoId)
+    if (!todoExists) {
+      return left(new TodoNotFoundError(todoId))
+    }
     const result = await this.todoRepository.delete(todoId)
     return right(result)
   }
